Add a mobile header nav
Choose to implement a simple custom burger menu after reviewing react
burger menu which didn't render as desired.

Adding always visible cta to header and removing react-visibility-sensor
logic which is to complex for the benefit it gives.

The visibility logic would have only been triggeerd to show the CTA in
the header on desktop when on the home page.